+++ NEWS.md
@@ -1,3 +1,11 @@
+## R package processR version 0.1.6
+===================================
+(2019-Apr-4)
+
+- New data teachers added
+
+- New function condPlot added
+
 ## R package processR version 0.1.5
 ===================================
 (2019-Mar-29)